Allow non RFC822 formatted emails
In the case that an email doesn't explicitly conform to RFC822 (in my case I ran into it in testing), the SendGrid gem will fail with an `undefined method []` because the match is expected to work. This instead just falls back to the originally given email. Behavior originally introduced in d2063823b0ec4c88cdb8e84997ebedafa9ef655f.