Update example to work in Rails console
Because `Mail` is used by Rails the "hello world" example doesn't work in the Rails console out of the box. This change fixes that and is also consistent with the existing reference to `SendGrid::API` which is also strictly speaking unnecessary if one is including this module in a class.

Perhaps there is another way to handle this but my view FWIW is that it's supremely important that simplest example work in the Rails console with no changes.