Remove unnecessary require statements
Because none of these files use the JSON library, it's not necessary
to use `require 'json'` at the beginning of the files. This is only a
minor cleanup but using more `require` statements does have the
potential to slow down applications when more gems are installed and
the include path grows within Ruby.