feat: adds rack middleware to make request verification easier in rack apps. (#428)
This adds a class that can be used to verify requests with a public key in rack applications like Rails or Sinatra.
This also changed the  class to fail verification if an error was thrown. In the case where a signature was missing, OpenSSL would throw an error instead of returning false for an invalid signature.