test: add Simplecov Local Enhancements (#368)
As a contributor, I'd like to know what baseline code coverage is and how my enhancements affect it.

It was actually fairly difficult to find a good location to insert the SimpleCov.start and keep the results consistent. But after several hours, the final state of this PR seems to be the best.

No documentation was updated because code coverage is now part of the default rake run, with no further action required.